Move notices to a separate file
